Hot fix to resolve breaking changes (#878)
* fix promptbehavior enum and not use extension methods

* rename file

* update to 3.17.1

* create new test certificate

* fix doc

* add obsolete tags

* move AdalError to separate file
