+++ devApps/AdalUniversalTestApp/AdalUniversalTestApp.Windows/AdalUniversalTestApp.Windows.csproj
@@ -20,7 +20,7 @@
     <SynthesizeLinkMetadata>true</SynthesizeLinkMetadata>
     <ProjectTypeGuids>{BC8A1FFA-BEE3-4634-8014-F334798102B3};{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}</ProjectTypeGuids>
     <PackageCertificateKeyFile>AdalUniversalTestApp.Windows_TemporaryKey.pfx</PackageCertificateKeyFile>
-    <PackageCertificateThumbprint>46F0D976A20E01CB3A7F005F387727D1A6C9AF58</PackageCertificateThumbprint>
+    <PackageCertificateThumbprint>605BAFB9C322BC40C67B938D2C0CF160838002C4</PackageCertificateThumbprint>
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
     <PlatformTarget>AnyCPU</PlatformTarget>
