+++ src/Microsoft.IdentityModel.Clients.ActiveDirectory/Features/NonWinCommon/PromptBehavior.cs
@@ -46,18 +46,18 @@ namespace Microsoft.IdentityModel.Clients.ActiveDirectory
         Always,
 
         /// <summary>
-        /// Prompt the user to select a user account even if there is a token that meets the requirements
-        /// already in the cache. This enables an user who has multiple accounts at the Authorization Server to select amongst 
-        /// the multiple accounts that they might have current sessions for. 
-        /// </summary>
-        SelectAccount,
-
-        /// <summary>
         /// Re-authorizes (through displaying webview) the resource usage, making sure that the resulting access
         /// token contains updated claims. If user logon cookies are available, the user will not be asked for 
         /// credentials again and the logon dialog will dismiss automatically.
         /// </summary>
-        RefreshSession
+        RefreshSession,
+
+        /// <summary>
+        /// Prompt the user to select a user account even if there is a token that meets the requirements
+        /// already in the cache. This enables an user who has multiple accounts at the Authorization Server to select amongst 
+        /// the multiple accounts that they might have current sessions for. 
+        /// </summary>
+        SelectAccount
 
         // NB: We do not support the "Never" option on Xamarin iOS or Android.
     }
